requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
requests
django-revproxy-0.10.0
python-stdnum=1.16
requests_mock=1.9.3
